Move body of RandomEvent to Event class.
