+++ src/app/backend/event/event.ts
@@ -1,14 +1,13 @@
-export class Event {}
-
-export class RandomEvent extends Event {
+export class Event {
   name: string;
   text: string;
   img: HTMLImageElement;
 
   constructor(name, text, img) {
-    super();
     this.name = name;
     this.text = text;
     this.img = img;
   }
 }
+
+export class RandomEvent extends Event {}
